### 2019-05-25 #309 @tyluRp
Submission: exoplanets
Submitting Author: Tyler Littlefield (@tyluRp)
Repository: https://github.com/tyluRp/exoplanets
Version submitted: 0.0.0.9000
Version accepted: TBD

---

```
Package: exoplanets
Title: Provides access to NASA's Exoplanet Archive, see <https://exoplanetarchive.ipac.caltech.edu/index.html>
Version: 0.0.0.9000
    person(given = "Tyler",
           family = "Littlefield",
           email = "tylurp1@gmail.com",
Description: The goal of exoplanets is to provide access to NASA's Exoplanet
  Archive database by supplying a table name or by writing the URL query
  manually. For more information please read the documentation
  <https://exoplanetarchive.ipac.caltech.edu/index.html>.
    covr,
    dplyr
    utils
```

     * The package falls under the data retrieval category because it pulls data from [NASA's Exoplanet Archive Database](https://exoplanetarchive.ipac.caltech.edu/docs/program_interfaces.html).

     * Target Audience: astronomers, space enthusiasts, NASA datanauts
     * Applications: Planet hunting

     * None that I am aware of, there is a [thread](https://github.com/ropensci/unconf17/issues/67) that discusses NASA APIs but I'm not sure if the Exoplanet Archive was ever targeted or implemented in an R package. Additionally there is [nasadata](https://github.com/Eflores89/nasadata) which provides access to NASA's EONET API. Same API page but doesn't include the Exoplanet archive as far as I can tell.

     * #308
