### 2019-06-10 #309 @geanders
### Editor checks:

---

#### Editor comments

This looks like a really interesting package, thanks for submitting!

We run a few standard checks on all new packages, using `spelling::spell_check_package()`, `goodpractice::gp()`, `devtools::spell_check()`, and `covr::package_coverage()`.

From `spelling::spell_check_package()` and `devtools::spell_check()`: No true spelling errors identified, all looks good for that.

From `covr::package_coverage()`: 100% coverage, great!

From `goodpractice::gp()`: There are several lines of code longer than 80 characters that could be shortened.

Here are some small fixes to consider based on those checks:

- [ ] Please add a 'Language' field to the 'DESCRIPTION' file
(e.g., 'Language = en-US'.
- [ ] Try to get the following lines of code under 80 characters:
in 'exoplanets.R', lines 37, 42, 45, 62, 65, and 67; in
'test-exoplanets.R', line 12.

For the next step, I'll assign reviewers for the submission. Please let me know if you have any questions during the process!

---

Reviewers:
Due date:
