### 2019-06-13 #309 @tyluRp
Awesome!

I've added `Language: en-US` to the DESCRIPTION file and updated the `codemeta.json` file.

 I have shortened a few of the > 80 lines. A few remain but they are mostly due to a really long API URL. I would prefer to retain that URL instead of `paste0()`ing or `glue()`ing it together in chunks. If this is against policy just let me know and I'll make sure nothing exceeds the 80 character limit 😸
