### 2019-06-19 #309 @geanders
@tyluRp : Thanks so much for those updates. I don't think we have hard-and-fast rules for anything in the packages, just guidance. If you have thought about good practice guidelines and have a case where it makes sense to you to make an exception, that seems reasonable.

We now have one reviewer assigned (@h21k) and I am continue to look for the second reviewer.
