### 2019-09-23 #309 @nmonhait
>##### For packages co-submitting to JOSS- **N/A**
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Hi @tyluRp, great work on this package! My comments are below:

**README**
- there is no statement about contributions or code of conduct (unless I am missing this somewhere/if so, this could be more explicit)
- I would like to see more about the data and who can use the package/how. You do a great job of this in this issues thread. I imagine if this were the first touchpoint for a user, they would appreciate some additional background
-it may be helpful to link to vignettes here

**General**
-there is some code that spans longer than 80 chars
-I had LaTex errors, this may be specific to my machine but might be something to look further into
-tell me more about this statement - "Finally, there are 3 tables which require an additional parameter in order to access them, these are all time series datasets:", why are these different?
-if I were a newcomer, I would need some elaboration on this sentence- "For example, the `kelttimeseries` table requires either the quarter or a Kepler ID, `exo("kelttimeseries&quarter=14")`. Note, these datasets are quite large and can take awhile. All of these additional parameter requirements are documented in the [documentation](https://exoplanetarchive.ipac.caltech.edu/docs/program_interfaces.html)."

**devtools::check()**

0 errors ✔ | 0 warnings ✔ | 0 notes ✔

**devtools::test()**

OK:       9

**goodpractice::gp()**

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide.
    Try make your lines shorter than 80 characters

    R/exoplanets.R:37:1
    R/exoplanets.R:65:1
    tests/testthat/test-exoplanets.R:12:1

**To summarize** I think you did a great job. My main suggestion is to add more detail aimed at the 'first-touchpoint' user. Thorough documentation and vignettes will allow newcomers to utilize the functionality of this package.
