### 2019-09-25 #309 @tyluRp
@nmonhait Thanks a bunch for all the feedback! I will start working on the things you have mentioned. I am curious about the LaTeX errors. If possible, could you elaborate? I do recall adding `exo_summary` recently and one of the attributes has a >= sign in unicode, this has since been removed but maybe there are some lurking around.

- [x] Add CONTRIBUTING.md
- [x] Add CODE_OF_CONDUCT.md
- [x] Add "getting started" vignette
- [ ] Add a clear and concise explanation of time series datasets and why they are different
- [ ] Add functionality to address time series datasets, perhaps a function `exo_timeseries` dedicated to these datasets

Let me know if I missed anything. Additionally, I recently made a _somewhat_ large change to the package. I've decided to use `httr::GET` instead of `read.csv` for the progress messages and to allow better handling of the requests for the future. For some reason, this change has caused the appveyor build to fail. If anyone is using Windows, I'd be interested to see if this package works for you. Additionally, if anyone knows why it's failing please reach out.

Thanks again for the feedback 👍
