### 2019-09-25 #309 @geanders
Thank you so much for your review, @nmonhait! @h21k, do you have an update on when you would be able to add your review?

@tyluRp : Regarding your Windows question, I'm using a Mac, so I can't check. Perhaps @nmonhait or @h21k is using Windows and could check. Also, have you tried out the package through http://win-builder.r-project.org/? I'm not sure if that's going to provide more details for you than the Appveyor output, but it might be worth trying if you haven't yet. If you're using the `devtools` package, the `build_win` and `check_win` functions both interface with the winbuilder site.
