### 2019-09-26 #309 @tyluRp
@geanders Thanks for the advice. I've just tried this out on a windows machine and I am getting the same error:

```r
Error in curl::curl_fetch_memory(url, handle = handle) :
  Failure when receiving data from the peer
```

I will use the `devtools` functions to see if this provides any additional information.
