### 2019-09-28 #309 @tyluRp
Wanted to provide an update for any Windows users. I filed an issue for the `curl::curl_fetch_memory` error here https://github.com/jeroen/curl/issues/206. Some interesting responses.

I've decided to add two utility functions to the package for handling requests differently depending on the operating system (windows, or anything else), you can see the additions here https://github.com/tyluRp/exoplanets/commit/6b93fb231cb408936be0d03fea42bb845b12e0c5. This has fixed the appveyor build.
