### 2019-11-03 #309 @h21k
Hi all,

apologies for the delay.
I should have mine ready very soon.

Best wishes,

Frank

On Sun, 3 Nov 2019 at 04:01, Tyler Littlefield <notifications@github.com>
wrote:

> Hi @nmonhait <https://github.com/nmonhait>, thanks again for your
> feedback. I've added a vignette with a walk through on how to use the
> package and access data from the archive. I've also added additional helper
> functions for the time series datasets, the intro vignette covers these new
> functions as well.
>
> Code of conduct and contributing guidelines have been added. When you have
> a moment, please let me know your thoughts.
>
