### 2020-07-23 #309 @h21k
Hi Tyler,

thank you for the info - I should be able to pick this up by end of the
month.
So happy to review.

Best wishes,

Frank

On Wed, 22 Jul 2020 at 21:59, Tyler Littlefield <notifications@github.com>
wrote:

> Hi @geanders <https://github.com/geanders> and @h21k
> <https://github.com/h21k>, I have someone (@mathidachuk
> <https://github.com/mathidachuk>) who is interested in reviewing the
> package if needed. Please let us know!
>
