### 2020-07-23 #309 @geanders
We would love to get @mathidachuk involved as an ROpenScie reviewer! For this specific package, we do have a conflict of interest policy. Let's start by having the two of you take a look at the [conflict of interest policies](https://devguide.ropensci.org/policies.html#coi). I think if you are colleagues at the same institution, then @mathidachuk might be conflicted out from this review. However, we have [some tips on becoming a reviewer](https://devguide.ropensci.org/softwarereviewintro.html#whyreview), which includes a link to a form for joining a database of reviewers that we use a lot to help find reviewers for specific packages. This would give @mathidachuk the chance to review other packages, even if there is a conflict of interest for this one.
