### 2020-10-05 #309 @maelle
:wave: @h21k, are you still able to review the package? If so here's the link to our [reviewers guide](https://devguide.ropensci.org/reviewerguide.html) and [review template](https://devguide.ropensci.org/reviewtemplate.html).
