### 2021-05-01 #309 @maelle
Thanks again for your submission and patience @tyluRp!
:warning: `exo_columns()` fails because the [data changed upstream](https://exoplanetarchive.ipac.caltech.edu/cgi-bin/nstedAPI/nph-nstedAPI?table=exoplanets&getDefaultColumns&format=csv).  Therefore the tests and vignette building fail. I've made some other comments.  Some of them are suggestions, the most important comments are about documentation and improving testing. Happy to discuss!

- [ ] **Examples** (that run successfully locally) for all exported functions

---

This is a nice package for people who'll need to use this data in R! :dizzy:

#### CRAN submission

I see traces of a CRAN submission, but the packages isn't on CRAN yet, what happened? In any case, now it'd make sense to wait until after the review changes to re-submit. Also note that there is no obligation to submit the package to CRAN.

* You could rename the vignette to "explanets.Rmd" for having it automatically listed [under "Get started" in the navbar by pkgdown](https://pkgdown.r-lib.org/reference/build_site.html#yaml-config-navbar).
* I saw two "click here" links, which is [better avoided](https://webaccess.berkeley.edu/ask-pecan/click-here) in particular for accessibility.
    * https://github.com/tyluRp/exoplanets/blob/6bff47e1c3693ee093e86b658eb3a1efd4c721d4/README.Rmd#L139
    * https://github.com/tyluRp/exoplanets/blob/6bff47e1c3693ee093e86b658eb3a1efd4c721d4/R/exo_summary.R#L4
* In the documentation of some functions you write e.g. "See the KELT page", it would be nice to have links to such pages.
* In the documentation of `exo_tables()` it'd be nice to have an example as it'd mean the table would be present in the pkgdown website.
* When documenting `exo_raw()` could you add some explanations of the example?
* In the docs of e.g. `exo()`, for the parameter `cols` you could add how to know what columns are default, what columns are available (simply linking to `exo_columns()` docs I suspect).
* The quiet argument, and maybe others, is repeatedly documented. You could use [roxygen2's parameter inheritance or templates](https://roxygen2.r-lib.org/articles/rd.html#do-repeat-yourself-1) to avoid this.
* The README image is quite cool!
    * Could you please add an [alternative text](https://blog.rstudio.com/2021/04/20/knitr-fig-alt/) for it?
    * Could you explain more about the image and how you created it using the package, perhaps in a vignette linked from the README?
* I'd recomment [precomputing the vignette](https://ropensci.org/blog/2019/12/08/precompute-vignettes/) (especially if you submit your package to CRAN).

#### Contributing information

There is no CONTRIBUTING.md at the moment.
We have some [basic guidance](https://devguide.ropensci.org/collaboration.html#contributing-guide) around this in the dev guide, that I will update in the upcoming weeks based on our recent [community call](https://ropensci.org/commcalls/apr2021-pkg-community/).

When time comes to transfer your repository to the ropensci organization, we'll ask you remove the Code of Conduct file as [rOpenSci Code of Conduct will apply to your package](https://devguide.ropensci.org/collaboration.html#coc-file).

* Why not use `httr::RETRY()` (or `httr::GET()`) for downloading the data? You could then [stop for status](https://books.ropensci.org/http-testing/graceful.html#user-facing-grace-how-your-package-actually-works).
* Would it make sense to try and add a progress bar for downloads? As they are failing for now, I do not know how long they take, though.
* For formatting the messages e.g. the ones with URLs you might be interested in the [cli package](https://cli.r-lib.org/).
* Why do you memoise the functions in `.onLoad()`? Genuinely curious, as we use a different strategy in e.g. [opencage](https://github.com/ropensci/opencage/search?q=memoise).

Tests currently fail, but I also find them a bit light at the moment.

* You could test more API calls. I'd advise to mock them somehow, using e.g. vcr or httptest. We updated the [HTTP testing in R book](https://books.ropensci.org/http-testing/introduction.html) last year. I'm happy to answer any question on the topic.
* You could test the user-friendly error messages, using [testthat snapshot testing](https://testthat.r-lib.org/articles/snapshotting.html).

#### Continuous integration

* I see you use GitHub Actions, just a note that tidyverse repos seem to be switching towards pak-based workflows (whose [names end with `-pak`](https://github.com/r-lib/actions/tree/master/examples)) instead of remotes-based workflows.
* To catch a problem as the one in `exo_tables()` you could add a [regularly scheduled workflow](https://books.ropensci.org/http-testing/real-requests-chapter.html#when-to-make-real-requests). Note that in GitHub Actions, after 60 days without repo activity, scheduled workflows are deactivated.

Thanks again for your submission & work on this neat&useful package! :ringed_planet:
