### 2021-05-01 #309 @tyluRp
Thanks for the review!
1. Regarding CRAN, there was an odd bug accessing the API from a windows machine in the check so I could never get the build to pass.
2. Regarding `httr`, I actually used this from the beginning but eventually removed it thinking this might somehow resolve the CRAN issue, https://github.com/ropensci/software-review/issues/309#issuecomment-534840574
3. When I had `httr` progress bar was a nice to have, I made it so you could turn it on/off but had since removed due to point above.
4. Regarding `memoise`, I just wanted to check it out and see how caching results work. It seems like it could be nice if someone is hitting the api more than once with the same query.

The 60 day scheduled check sounds really nice, especially for a package that depends on an API. I'll review this feedback over the weekend and start updating things. Thanks!
