### 2021-05-02 #309 @tyluRp
@maelle Okay! So I've gone ahead an basically rewrote the entire package due to API changes. AFAIK, the function used to get column information broke last Friday so basically everything broke as you mentioned. This rewrite uses the new "TAP" service offered by the exoplanet archive. I've decided to support the new interface for now as it looks like tables from the old interface are planned to be migrated to "TAP" in the future. If things settle down, I may decide to support the old API.

Summary of changes:

1. Entire rewrite, supporting new "TAP" service instead of old "API" service
2. Precomputed/updated vignette
3. Use of `httr` for requests instead of `read.csv`
4. Progress bars
5. Updated table information dataset (with example)
6. Added contributing guidelines
7. Update workflows to use `-pak` based
8. Added alternative text to README
9. Removed "click here" style links

Pending items:

1. Tests! I've got to read up on this and add some tests for the actually requests.

Anyway, this should be good enough to actually see the package in action. I will post an update once I'm happy with the tests. Please let me know if you have any questions.
