### 2021-05-02 #309 @tyluRp
Alright, I've added:

1. Tests with `httptest`
2. [Weekly r cmd check](https://github.com/tyluRp/exoplanets/blob/master/.github/workflows/check-pak-cron.yaml#L4-L6), basically the check-pak.yml workflow but triggered by cron instead of commits (I assume this is how it's done, never tried before!)

Also wanted to say thanks a bunch, this feedback has taught me quite a bit, specifically:
1. How awesome `httptest` is, I'm hoping this makes my life easier when trying to submit this to CRAN.
2. Precompiled vignettes! Ditto on point 1, I hope it makes submission easier.
3. Reoccurring workflows!
