### 2021-05-03 #309 @maelle
Thanks @tyluRp for all your work on this! :rocket: :sparkles:

My answers

* Please summarize changes in NEWS.md especially as the changes were quite big. :slightly_smiling_face:
* I feel less bad about the delay now that I know it means I reviewed the package right after the API calls. :grin: (seriously, thanks again for your patience!)
* Regarding CRAN it's good that now your package has pre-computed vignettes, not-run examples and mocked tests as it means no API call will be made from CRAN. :grin:
* Regarding memoise, I realize [`.onLoad()` is actually used in opencage](https://github.com/ropensci/opencage/blob/main/R/zzz.R) so my comment was not valid.
* A very small suggested tweak, if you run `desc::desc_normalize()` dependencies will be alphabetically ordered and Imports will come before Suggests which is more usual.
* Regarding your wondering whether things will settle down with the API, do you need to contact the API maintainers?
* The URL in https://github.com/tyluRp/exoplanets/blob/577260b3ce89f1e43e7bf781ffaed6ace1b9222a/R/data.R#L10 needs to be wrapped in `\url{}` to be clickable on the pkgdown website in particular.
* Wouldn't it make sense to use [`httr::RETRY()` instead of `httr::GET()`](https://blog.r-hub.io/2020/04/07/retry-wheel/)?
* In the scheduled GitHub Actions run you need to make sure httptest is turned off (at least for one of the builds) as otherwise you can't catch errors due to changes in the API. I don't remember how to turn off httptest, whether you need to delete mock files or whether you can add some sort of switch based on an environmental variable.
* In the vignette when re-recreating the plot why not add a few sentences about why to plot these things and why to label "TRAPPIST-1 e"? I.e. providing some context for the use case.

Thanks again!
