### 2021-05-03 #309 @maelle
One last question actually, why keep the random aspect of the tests and not add tests corresponding to the same use cases as the examples in the manual?
