### 2021-05-04 #309 @tyluRp
* NEWS.md updated.
* `desc::desc_noarmalize()` used.
* I do not think I need to reach out the maintainers at this time. Will just need to review and monitor their changelog from time to time.
* Wrapped link in `\url{}`.
* Using `httr::RETRY()`.
* Removed scheduled github action, will take a look at this at a later time. Note to self, I'll want to take a look at https://github.com/nealrichardson/httptest/issues/40.
* Added some text on why "TRAPPIST-1 e" is labeled.
* Added tests for most of the examples (excluded the one that requests the default `ps` table which is huge). As for why they are random, I was just trying to cover things outside the examples, like expected errors, different parameters/scenarios, etc.

Please let me know if theres anything else I missed!
