### 2021-05-04 #309 @tyluRp
By the way, the `desc` package is pretty cool and when I used it, it showed the following for the authors part of the DESCRIPTION:
```
Authors@R (parsed):
  * Tyler Littlefield <tylerlittlefield@hey.com> [aut, cre] (<https://orcid.org/0000-0002-6020-1125>)
```
But `devtools::document()` was not liking it. Am I supposed to be able to record authors in the DESCRIPTION file this way?
