### 2021-05-05 #309 @maelle
Some comments (probably last comments). Thanks for your work!

* I've opened a [PR to fix the link of the GitHub Actions badge](https://github.com/tyluRp/exoplanets/pull/4).
* Instead of having to force silence of the package why not add a verbose option / argument? Having an option for verbosity is e.g. the approach in [usethis](https://usethis.r-lib.org/articles/articles/usethis-setup.html#store-default-values-for-description-fields-and-other-preferences-1) and [gargle](https://github.com/r-lib/gargle/search?q=verbosity).
* When you use `"data.frame" %in% class` in expectations in tests it'd be more elegant to use [`expect_type()`](https://testthat.r-lib.org/reference/inheritance-expectations.html) instead.
* Instead of `expect_error()` if you use [`expect_snapshot_error()`](https://testthat.r-lib.org/articles/snapshotting.html#other-types-of-snapshot-1) your test will be more precise.
* For these two last tips maybe you need to switch to testthat 3d edition in which case you can refer to [its release post](https://www.tidyverse.org/blog/2020/10/testthat-3-0-0/).
* Reg desc what it prints in the console is not exactly what needs to go in the file, but when you run `desc::desc_normalize()` desc itself edits the DESCRIPTION file. Not sure if I understood the question right though!
