### 2021-05-05 #309 @tyluRp
* Thanks for the PR! I've added you to the description file as a contributer/reviewer.
* A quiet option sounds nice but I'm also using the `quiet` function in the `testthat` script to avoid this message "Using redact.R in 'exoplanets'" message that seems to appear when testing. I want the test log to be as clean as possible.
* Thanks for the suggestion on `expect_type`, I went with `expect_s3_class` because `typeof` on a dataframe is list and I'll need to be able to distinguish lists vs dataframes.
* Using `expect_snapshot_error`, thanks for the suggestion!

Please let me know if there is anything else :)
