### 2021-05-06 #309 @tyluRp
@maelle I think I got most of it except:
1. The redirect, wasn't really sure what I am supposed to do but I see that the ropensci docs are live! https://docs.ropensci.org/exoplanets/
2. General admin access, I can't seem to update the documentation URL in the github interface
![image](https://user-images.githubusercontent.com/35909636/117366335-71e26080-ae75-11eb-8d4b-b4ac3dc9cf51.png)
