### 2021-05-07 #309 @maelle
@tyluRp

* I've re-added your admin rights to the repo.
* I'm interested in knowing what to add to https://devguide.ropensci.org/redirect.html#github-pages to clarify it so could you please tell me your questions? Thank you!
